move print-* —→ print/*
